update all archive and quick edit icons
